Updated the UI to include a track list and now playing tab.
Cleaned up a few things in the engine:
Instead of trying to play a different group of tracks at the start of each iteration, it now constantly plays all tracks but fades the tracks in and out.

TODO: Add a few more songs with different styles and tempos.
Add a few songs with SOME tags that match other songs, but not a complete match. This way I can start working on the song selection logic
Add a bit in the scene configuration screens to allow music tags to be set, and then use those tags to select songs.
Integrate with the combat tracker
Integrate with the collective HP of the party to dynamically adjust music intensity based on the group's current health status.